# docs/swagger.yaml
...
info:
  contact: {}
paths:
  /api/excel/export:
    get:
      consumes:
      - application/json
      description: 数据导出
      parameters:
      - description: 当前第几页，默认1
        in: query
        name: cur_page
        type: integer
      - description: 一页显示数据量，默认20
        in: query
        name: page_size
        type: integer
      - description: 导出文件类型(csv,xlxs)
        in: query
        name: file_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 文件内容字节流
          schema:
            items:
              type: integer
            type: array
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  type: string
                errors:
                  type: string
                "no":
                  type: integer
              type: object
      summary: 数据导出
      tags:
      - Excel操作
  /api/excel/import:
    post:
      consumes:
      - application/json
      description: 数据导入
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  type: string
                msg:
                  type: string
                "no":
                  type: integer
              type: object
      summary: 数据导入
      tags:
      - Excel操作
...
swagger: "2.0"